+++ packages/create-thrashplay-app/_templates/app/init/2.0_license/LICENSE.ejs.t
@@ -1,3 +1,3 @@
 ---
-sh: "<% if (licenseId) { %>node <%= scriptsDir %>/fetch-license.js <%= licenseId %> > <%= projectDir %>/LICENSE <% } else { %>echo No license specified, skipping LICENSE generation.<% } %>"
+sh: "<% if (licenseId) { %>node <%= createThrashplayAppScriptsDir %>/fetch-license.js <%= licenseId %> > <%= projectDir %>/LICENSE <% } else { %>echo No license specified, skipping LICENSE generation.<% } %>"
 ---