+++ packages/create-thrashplay-app/src/templates/lerna/add-package/0.0_base/package.json.ejs.t
@@ -4,6 +4,9 @@ to: <%= projectDir %>/packages/<%= name %>/package.json
 {
   "name": "<%= name %>",
   "version": "0.0.0",
+  "publishConfig": {
+    "access": "public"
+  },
   "dependencies": {
   },
   "devDependencies": {